feat: introduced AdapterException class
