+++ aidial_adapter_openai/utils/adapter_exception.py
@@ -1,21 +1,22 @@
 import json
-from typing import Any, MutableMapping
+from typing import Any, Dict
 
 from aidial_sdk.exceptions import HTTPException as DialException
+from fastapi import HTTPException as FastAPIException
 from fastapi.responses import Response as FastAPIResponse
 
 
 class ResponseWrapper(Exception):
     content: Any
     status_code: int
-    headers: MutableMapping[str, str] | None
+    headers: Dict[str, str] | None
 
     def __init__(
         self,
         *,
         content: Any,
         status_code: int,
-        headers: MutableMapping[str, str] | None,
+        headers: Dict[str, str] | None,
     ) -> None:
         super().__init__(str(content))
         self.content = content
@@ -38,6 +39,13 @@ class ResponseWrapper(Exception):
             headers=self.headers,
         )
 
+    def to_fastapi_exception(self) -> FastAPIException:
+        return FastAPIException(
+            status_code=self.status_code,
+            detail=self.content,
+            headers=self.headers,
+        )
+
     def json_error(self) -> dict:
         return {
             "error": {
@@ -51,7 +59,7 @@ AdapterException = ResponseWrapper | DialException
 
 
 def _parse_dial_exception(
-    *, status_code: int, headers: MutableMapping[str, str], content: Any
+    *, status_code: int, headers: Dict[str, str], content: Any
 ) -> DialException | None:
     if isinstance(content, str):
         try:
@@ -60,11 +68,6 @@ def _parse_dial_exception(
             return None
     else:
         obj = content
-
-    # The content length is invalidated as soon as
-    # the original content is lost
-    if "Content-Length" in headers:
-        del headers["Content-Length"]
 
     if (
         isinstance(obj, dict)
@@ -84,14 +87,14 @@ def _parse_dial_exception(
             param=param,
             code=code,
             display_message=display_message,
-            headers=dict(headers.items()),
+            headers=headers,
         )
 
     return None
 
 
 def parse_adapter_exception(
-    *, status_code: int, headers: MutableMapping[str, str], content: Any
+    *, status_code: int, headers: Dict[str, str], content: Any
 ) -> AdapterException:
     return _parse_dial_exception(
         status_code=status_code, headers=headers, content=content
